Trait rp6::avr::device::pin::Pin

pub trait Pin {
    type DDR: Register<T = u8>;
    type PORT: Register<T = u8>;
    type PIN: Register<T = u8>;

    const OFFSET: u8;
    const MASK: u8;

    fn set_direction(direction: DataDirection) { ... }
    fn set_input() { ... }
    fn set_output() { ... }
    fn set_high() { ... }
    fn set_low() { ... }
    fn toggle() { ... }
    fn is_high() -> bool { ... }
    fn is_low() -> bool { ... }
}

An IO pin.

Required Associated Types

type DDR: Register<T = u8>

The associated data direction register.

type PORT: Register<T = u8>

The associated port register.

type PIN: Register<T = u8>

Reads from the register will read input bits.

Required Associated Constants

const OFFSET: u8

The numeric offset of the Pin in the register

const MASK: u8

The mask of the pin used for accessing registers.

Provided Methods

fn set_direction(direction: DataDirection)

Sets the data direction of the pin.

fn set_input()

Sets the pin up as an input.

fn set_output()

Sets the pin up as an output.

fn set_high()

Set the pin to high.

The pin must be configured as an output.

fn set_low()

Set the pin to low.

The pin must be configured as an output.

fn toggle()

Toggles the pin.

The pin must be configured as an output.

fn is_high() -> bool

Check if the pin is currently high.

The pin must be configured as an input.

fn is_low() -> bool

Checks if the pin is currently low.

The pin must be configured as an input.