Struct rp6::avr::interrupt::mutex::DynamicMutex

pub struct DynamicMutex<T: ?Sized> {
    data: UnsafeCell<RefCell<T>>,
}

DynamicMutex implements a guard that is safe for “concurrent” data access on single-core devices that can only experience non-atomic data access via interrupts. It restricts data access to within a CriticalSection, which can only be obtained within a closure given to without_interrupts.

The data is wrapped in a core::cell::RefCell for mutability, which is dynamically checked for consistency and can lead to a panic. Whenever possible, it is recommended to use a Mutex instead, which is based on a Cell.

Fields
data: UnsafeCell<RefCell<T>>
Implementations

impl<T: Sized> DynamicMutex<T>

pub const fn new(data: T) -> Self

Create a new DynamicMutex wrapping data of type T.

impl<T: ?Sized> DynamicMutex<T>

pub fn lock<'access_time>(
&'access_time self,
_cs: &'access_time CriticalSection
) -> &'access_time RefCell<T>

Get uninterrupted access to the encapsulated data. Can only be called inside a CriticalSection, ensured by requiring a reference to one.