Module rp6::avr::legacy::serial

A serial.

WARNING The current implementation of this will only work on ATmega328.

Modules

hack 🔒

Structs

Serial

A serial connection. WARNING The current implementation of this will only work on ATmega328.

Enums

CharacterSize

Mode

Parity

StopBits

Functions

do_read 🔒

do_write 🔒

ready_to_receive

ready_to_transmit

receive

Does a blocking read of one byte

transmit

Does a blocking transfer of one byte

try_receive

try_transmit