Module rp6::uart

UART = “Universal Aynchronous Receiver Transceiver”

This module contains data transfer functions that allow easy access to the robot’s serial UART connection. Receiving messages is asynchronous (using interrupts) and uses ringbuffers for intermediate storage of the received messages.

Modules

serial_writable 🔒

Module that implements Serial::write and formatting behavior for types.

Macros

uwrite

Structs

Serial

Struct managing all access to the robot’s serial port connection

String

A fixed capacity String

Constants

BAUD_HIGH

BAUD_LOW

Define constants for RP6 baudrates.

UBRR_BAUD_HIGH

UBRR_BAUD_LOW

Traits

SerialWritable

Trait to allow implementing specific Serial::write behavior for types.

SerialWritableDecimal

Trait to allow implementing specific Serial::write_dec behavior for types.

SerialWritableHexadecimal

Trait to allow implementing specific Serial::write_hex behavior for types.

StringType

Trait to allow instantiation and passing as &str for a type.

uDisplay

Just like core::fmt::Display

uWrite

A collection of methods that are required / used to format a message into a stream.

Functions

size_of

Returns the size of a type in bytes.