Struct rp6::uart::Serial

pub struct Serial;

Struct managing all access to the robot’s serial port connection

Implementations

impl Serial

pub fn init()

Initialize the serial connection on pins RX and TX.

pub fn enable_USART_RXC_interrupt()

Enable the USART_RXC interrupt

pub fn enable_USART_TXC_interrupt()

Enable the USART_TXC interrupt

pub fn disable_USART_RXC_interrupt()

Disable the USART_RXC interrupt

pub fn disable_USART_TXC_interrupt()

Disable the USART_TXC interrupt

pub fn set_baudrate_low()

Configure serial connection to low baudrate UBRR_BAUD_LOW.

pub fn set_baudrate_high()

Configure serial connection to high baudrate UBRR_BAUD_HIGH.

pub fn read_raw() -> u8

Reads a single raw byte from the Serial connection. Blocks until the processor is ready to receive the next byte, i.e., the corresponding bit UDRE is set in UCSRA.

pub fn write_raw(b: u8)

Writes a single raw byte to the Serial connection. Blocks until the processor is ready to send the next byte, i.e., the corresponding bit UDRE is set in UCSRA.

pub fn write<T: SerialWritable>(value: T)

Write something to the Serial connection. By default, supports &str, char, and basic number types (in decimal notation).

pub fn write_dec<T: SerialWritableDecimal>(value: T)

Write a number formatted as decimal to the Serial connection.

pub fn write_hex<T: SerialWritableHexadecimal>(value: T)

Write a number formatted as hexadecimal to the Serial connection.

pub fn new_line()

Write a '\n' (newline character) to the serial connection.